// src/app/models/admin-patient.ts
export interface Patient {
    id: string;
    email: string;
    first_name: string;
    last_name: string;
    phone_number: string;
    address: string;
    gender: string;
    nss: string;
    date_of_birth: string;
    place_of_birth: string;
    emergency_contact_name: string;
    emergency_contact_phone: string;
    medical_condition: string;
    profile_picture: string;
}